requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.1.1
Automat==20.2.0
black==20.8b1
certifi==2020.12.5
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
defusedxml==0.6.0
distlib==0.3.1
dj-static==0.0.6
Django==3.1.4
django-allauth==0.44.0
django-cors-headers==3.6.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.12.2
-e git+https://github.com/mobolic/facebook-sdk.git@3fa89fec6a20dd070ccf57968c6f89256f237f54#egg=facebook_sdk
filelock==3.0.12
flake8==3.8.4
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.3.0
incremental==17.5.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.22
mysqlclient==2.0.2
oauthlib==3.1.0
pathspec==0.8.1
Pillow==8.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.4.2
pylint-plugin-utils==0.6
pyOpenSSL==20.0.1
python3-openid==3.2.0
pytube==10.4.1
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
static3==0.7.0
stripe==2.55.1
toml==0.10.2
twilio==6.51.0
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
virtualenv==20.2.2
wrapt==1.12.1
zipp==3.4.0
zope.interface==5.2.0
